Added Party class
Created a class for the user to build their team of Pokémon. 
With a function to add Pokémon to the team if the team has less than 6 Pokémon.
Also added the function to swap a Pokemon to the front of the list for battle and putting the one who was in the front in its spot.